requirements.txt
django
psycopg2-binary
djangorestframework
Pillow
django-filter
django-cors-headers